Restructure composer and namespaces part I
