+++ composer.json
@@ -9,15 +9,12 @@
   },
    "autoload": {
     "files": [
-        "includes/Main.php",
-        "includes/Backend.php",
-        "includes/Frontend.php",
-        "includes/WooCommerce.php",
-        "includes/Gutenberg.php",
-        "includes/Functions.php",
-        "includes/emails/Emails.php"
-      ]
-    },
+        "functions.php"
+    ],
+    "psr-4": {
+        "WPFrontendDeleteAccount\\": "src/"
+    }
+  },
   "scripts": {
     "phpcs": [
       "phpcs -s -p"
