+++ functions.php
@@ -2,9 +2,11 @@
 /**
  * Core Functions of the plugin for both frontend and backend.
  *
- * @since  1.0.0
+ * @since 1.0.0
  *
  * @since 1.2.0 Rename filename from functions-wp-frontend-delete-account.php to Functions.php
+ *
+ * @since 1.3.0 Move from includes/ dir to root dir.
  */
 
 /**
