+++ src/Gutenberg.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WP_Frontend_Delete_Account;
+namespace WPFrontendDeleteAccount;
 
 /**
  * @since  1.0.0
@@ -55,5 +55,3 @@ class Gutenberg {
 		);
 	}
 }
-
-new \WP_Frontend_Delete_Account\Gutenberg();
