+++ src/emails/Emails.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WP_Frontend_Delete_Account;
+namespace WPFrontendDeleteAccount\Emails;
 
 /**
  * @since  1.2.0
